+++ src/Rocket.php
@@ -9,6 +9,7 @@ use JsonSerializable as JsonSerializableInterface;
 use Psr\Http\Message\MessageInterface;
 use Psr\Http\Message\RequestInterface;
 use Yansongda\Pay\Contract\PackerInterface;
+use Yansongda\Pay\Contract\ParserInterface;
 use Yansongda\Supports\Collection;
 use Yansongda\Supports\Traits\Accessable;
 use Yansongda\Supports\Traits\Arrayable;
@@ -28,7 +29,7 @@ class Rocket implements JsonSerializableInterface, ArrayAccess
 
     private string $packer = PackerInterface::class;
 
-    private ?string $direction = null;
+    private string $direction = ParserInterface::class;
 
     /**
      * @var \Yansongda\Supports\Collection|\Psr\Http\Message\MessageInterface|array|null
@@ -91,30 +92,24 @@ class Rocket implements JsonSerializableInterface, ArrayAccess
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getPacker(): string
     {
         return $this->packer;
     }
 
-    /**
-     * @param string $packer
-     * @return Rocket
-     */
     public function setPacker(string $packer): Rocket
     {
         $this->packer = $packer;
+
         return $this;
     }
 
-    public function getDirection(): ?string
+    public function getDirection(): string
     {
         return $this->direction;
     }
 
-    public function setDirection(?string $direction): Rocket
+    public function setDirection(string $direction): Rocket
     {
         $this->direction = $direction;