+++ src/Service/WechatServiceProvider.php
@@ -4,9 +4,7 @@ declare(strict_types=1);
 
 namespace Yansongda\Pay\Service;
 
-use Yansongda\Pay\Contract\ParserInterface;
 use Yansongda\Pay\Contract\ServiceProviderInterface;
-use Yansongda\Pay\Parser\CollectionParser;
 use Yansongda\Pay\Pay;
 use Yansongda\Pay\Provider\Wechat;
 
@@ -19,7 +17,6 @@ class WechatServiceProvider implements ServiceProviderInterface
     {
         $service = new Wechat();
 
-        Pay::set(ParserInterface::class, CollectionParser::class);
         Pay::set(Wechat::class, $service);
         Pay::set('wechat', $service);
     }